+++ src/main/java/com/example/todomono/repository/CustomerRepositoryInterface.java
@@ -3,7 +3,7 @@ package com.example.todomono.repository;
 import com.example.todomono.entity.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepositoryInterface extends JpaRepository<Customer, Integer> {
+public interface CustomerRepositoryInterface extends JpaRepository<Customer, Long> {
 
     Customer findByName(String name);
